move solc version to another crate
