+++ solx-standard-json/Cargo.toml
@@ -14,6 +14,7 @@ anyhow = "=1.0.89"
 rayon = "=1.10.0"
 serde = { version = "=1.0.210", "features" = [ "derive" ] }
 serde_json = { version = "=1.0.128", features = [ "arbitrary_precision" ] }
+semver = { version = "=1.0.23", features = [ "serde" ] }
 num = "=0.4.3"
 
 era-compiler-llvm-context = { git = "https://github.com/matter-labs/era-compiler-llvm-context", branch = "evm-llvmit-tests" }
